refactor: lift simulation logic into separate module
- Extracted the main lift simulation logic into .
- Implemented  and  functions to handle lift requests.
- Restored functionality for lifts to move and open/close doors correctly.
- Updated  to utilize the new lift simulation logic and manage pending requests effectively.
- Ensured that error handling for invalid input values is still in place.

This refactoring improves code organization and maintains existing lift functionality.
