+++ src/App.tsx
@@ -1,14 +1,7 @@
 import React, { useState, useEffect } from "react";
 import Floor from "./components/Floor";
 import Lift from "./components/Lift";
-
-type LiftType = {
-  id: number;
-  currentFloor: number;
-  isMoving: boolean;
-  doorsOpen: boolean;
-  style: React.CSSProperties;
-};
+import { useLiftSimulation, LiftType } from "./liftSimulation"; // Import the new lift logic
 
 const App: React.FC = () => {
   const [floorsCount, setFloorsCount] = useState<number>(0);
@@ -21,7 +14,6 @@ const App: React.FC = () => {
     Record<number, { up: boolean; down: boolean }>
   >({});
   const [errorMessage, setErrorMessage] = useState<string | null>(null); // State for error message
-  const floorHeight = 120; // Adjusting height for better visibility
 
   // Initialize disabled buttons state for each floor
   useEffect(() => {
@@ -31,33 +23,18 @@ const App: React.FC = () => {
       initialDisabledState[i] = { up: false, down: false };
     }
     setDisabledButtons(initialDisabledState);
-    console.log("Initialized disabled buttons state:", initialDisabledState);
   }, [floorsCount]);
 
-  useEffect(() => {
-    if (isSimulationStarted) {
-      initializeLifts(liftsCount);
-    }
-  }, [isSimulationStarted, liftsCount]);
-
-  const initializeLifts = (count: number) => {
-    const initialLifts: LiftType[] = [];
-    for (let i = 0; i < count; i++) {
-      initialLifts.push({
-        id: i + 1,
-        currentFloor: 0, // Set lifts to start at the ground floor
-        isMoving: false,
-        doorsOpen: false,
-        style: { bottom: "0px" }, // Ensure bottom is explicitly set to 0
-      });
-    }
-    setLifts(initialLifts);
-    console.log("Initialized lifts:", initialLifts);
-  };
+  const { findNearestAvailableLift, moveLift } = useLiftSimulation(
+    lifts,
+    pendingRequests,
+    setLifts,
+    setDisabledButtons,
+    setPendingRequests
+  );
 
   const handleStartSimulation = (e: React.FormEvent) => {
     e.preventDefault();
-    console.log("Starting simulation with:", { floorsCount, liftsCount });
 
     if (
       floorsCount <= 0 ||
@@ -65,17 +42,18 @@ const App: React.FC = () => {
       liftsCount <= 0 ||
       liftsCount > 10
     ) {
-      setErrorMessage("Please enter valid values (Floors: 1-100, Lifts: 1-10)"); // Set error message
+      setErrorMessage("Please enter valid values (Floors: 1-100, Lifts: 1-10)");
       return;
     }
 
     setErrorMessage(null); // Clear any previous error messages
     setIsSimulationStarted(true);
-    console.log("Simulation started.");
+    // Initialize lifts when simulation starts
+    const initializedLifts = initializeLifts(liftsCount);
+    setLifts(initializedLifts);
   };
 
   const handleResetSimulation = () => {
-    console.log("Resetting simulation...");
     setIsSimulationStarted(false);
     setLifts([]);
     setPendingRequests([]);
@@ -85,17 +63,9 @@ const App: React.FC = () => {
   };
 
   const handleLiftRequest = (floorNumber: number, direction: "up" | "down") => {
-    console.log(
-      `Lift request received for floor ${floorNumber} to go ${direction}.`
-    );
     const availableLift = findNearestAvailableLift(floorNumber);
-    console.log("Available Lift:", availableLift); // Log the available lift
-
     if (availableLift) {
-      moveLift(availableLift, floorNumber);
-      console.log(`Moving Lift ${availableLift.id} to Floor ${floorNumber}`); // Log when the lift is moving
-
-      // Disable the button that was pressed
+      moveLift(availableLift, floorNumber); // Call moveLift for the available lift
       setDisabledButtons((prev) => ({
         ...prev,
         [floorNumber]: {
@@ -105,99 +75,7 @@ const App: React.FC = () => {
       }));
     } else {
       setPendingRequests((prev) => [...prev, floorNumber]);
-      console.log(
-        `No available lifts. Added floor ${floorNumber} to pending requests.`
-      );
     }
-  };
-
-  const findNearestAvailableLift = (floorNumber: number): LiftType | null => {
-    let nearestLift: LiftType | null = null;
-    let minDistance = Infinity;
-    for (const lift of lifts) {
-      if (!lift.isMoving && !lift.doorsOpen) {
-        // Exclude lifts that are moving or have open doors
-        const distance = Math.abs(lift.currentFloor - floorNumber);
-        if (distance < minDistance) {
-          minDistance = distance;
-          nearestLift = lift;
-        }
-      }
-    }
-    console.log(
-      `Nearest available lift for floor ${floorNumber}:`,
-      nearestLift
-    );
-    return nearestLift;
-  };
-
-  const moveLift = (lift: LiftType, destinationFloor: number) => {
-    const floorsToMove = Math.abs(lift.currentFloor - destinationFloor);
-    const timePerFloor = 2000; // 2 seconds per floor
-    const totalTime = floorsToMove * timePerFloor;
-    const destinationBottom = `${destinationFloor * floorHeight}px`;
-
-    // Move the lift to the destination
-    console.log(
-      `Moving lift ${lift.id} to destination floor ${destinationFloor} in ${totalTime} ms.`
-    );
-    setLifts((prevLifts) =>
-      prevLifts.map((l) =>
-        l.id === lift.id
-          ? {
-              ...l,
-              isMoving: true,
-              style: {
-                ...l.style,
-                bottom: destinationBottom,
-                transition: `bottom ${totalTime}ms linear`,
-              },
-            }
-          : l
-      )
-    );
-
-    // After the lift reaches the destination, open the doors
-    setTimeout(() => {
-      console.log(
-        `Lift ${lift.id} has reached floor ${destinationFloor}. Opening doors.`
-      );
-      setLifts((prevLifts) =>
-        prevLifts.map((l) =>
-          l.id === lift.id
-            ? {
-                ...l,
-                isMoving: false,
-                doorsOpen: true,
-                currentFloor: destinationFloor, // Correctly update currentFloor here
-              }
-            : l
-        )
-      );
-
-      // Close the doors after 2.5 seconds
-      setTimeout(() => {
-        console.log(`Closing doors for lift ${lift.id}.`);
-        setLifts((prevLifts) =>
-          prevLifts.map((l) =>
-            l.id === lift.id ? { ...l, doorsOpen: false } : l
-          )
-        );
-
-        // Re-enable the buttons after the lift reaches the floor
-        setDisabledButtons((prev) => ({
-          ...prev,
-          [destinationFloor]: { up: false, down: false },
-        }));
-
-        // Handle pending requests if any
-        if (pendingRequests.length > 0) {
-          const nextFloor = pendingRequests.shift()!;
-          console.log(`Moving to next pending request for floor ${nextFloor}.`);
-          moveLift(lift, nextFloor);
-        }
-      }, 2500);
-    }, totalTime);
   };
 
   return (
@@ -207,7 +85,7 @@ const App: React.FC = () => {
         <button
           className="bg-red-600 text-white py-2 px-4 rounded mt-4"
           onClick={handleResetSimulation}
-          style={{ display: isSimulationStarted ? "block" : "none" }} // Show reset button only after starting simulation
+          style={{ display: isSimulationStarted ? "block" : "none" }}
         >
           Reset Simulation
         </button>
@@ -222,8 +100,7 @@ const App: React.FC = () => {
               className={`text-red-600 border border-red-600 p-2 rounded mb-4 ${
                 errorMessage ? "" : "hidden"
               }`}
-              role="alert" // Add role for accessibility
-              aria-live="assertive" // Add aria-live for dynamic updates
+              role="alert"
             >
               {errorMessage}
             </div>
@@ -234,11 +111,7 @@ const App: React.FC = () => {
               type="number"
               id="floorsCount"
               value={floorsCount}
-              onChange={(e) => {
-                const value = Number(e.target.value);
-                console.log("Updated floors count:", value);
-                setFloorsCount(value);
-              }}
+              onChange={(e) => setFloorsCount(Number(e.target.value))}
               min="1"
               max="100"
               required
@@ -253,11 +126,7 @@ const App: React.FC = () => {
               type="number"
               id="liftsCount"
               value={liftsCount}
-              onChange={(e) => {
-                const value = Number(e.target.value);
-                console.log("Updated lifts count:", value);
-                setLiftsCount(value);
-              }}
+              onChange={(e) => setLiftsCount(Number(e.target.value))}
               min="1"
               max="10"
               required
@@ -280,7 +149,7 @@ const App: React.FC = () => {
                 <Floor
                   key={index}
                   floorNumber={floorsCount - index}
-                  floorHeight={floorHeight}
+                  floorHeight={120} // Pass the defined height
                   handleLiftRequest={handleLiftRequest}
                   isTopFloor={floorsCount - index === floorsCount}
                   isGroundFloor={floorsCount - index === 0}
@@ -301,7 +170,7 @@ const App: React.FC = () => {
                     id={lift.id}
                     liftStyle={lift.style}
                     doorsOpen={lift.doorsOpen}
-                    currentFloor={lift.currentFloor} // Pass the current floor of the lift
+                    currentFloor={lift.currentFloor}
                   />
                 ))}
               </div>
@@ -311,6 +180,20 @@ const App: React.FC = () => {
       )}
     </div>
   );
+};
+
+const initializeLifts = (count: number): LiftType[] => {
+  const initialLifts: LiftType[] = [];
+  for (let i = 0; i < count; i++) {
+    initialLifts.push({
+      id: i + 1,
+      currentFloor: 0, // Set lifts to start at the ground floor
+      isMoving: false,
+      doorsOpen: false,
+      style: { bottom: "0px" }, // Ensure bottom is explicitly set to 0
+    });
+  }
+  return initialLifts;
 };
 
 export default App;
